Hscv {ks}	R Documentation

Smoothed cross-validation (SCV) bandwidth selector

Description

SCV bandwidth for 1- to 6-dimensional data.

Usage

Hscv(x, nstage=2, pre="sphere", pilot, Hstart, binned, 
     bgridsize, amise=FALSE, deriv.order=0, verbose=FALSE, optim.fun="optim")
Hscv.diag(x, nstage=2, pre="scale", pilot, Hstart, binned, 
     bgridsize, amise=FALSE, deriv.order=0, verbose=FALSE, optim.fun="optim")
hscv(x, nstage=2, binned=TRUE, bgridsize, plot=FALSE)

Arguments

x
vector or matrix of data values

pre
"scale" = pre.scale, "sphere" = pre.sphere

pilot
"amse" = AMSE pilot bandwidths
"samse" = single SAMSE pilot bandwidth
"unconstr" = single unconstrained pilot bandwidth
"dscalar" = single pilot bandwidth for deriv.order>0
"dunconstr" = single unconstrained pilot bandwidth for deriv.order>0

Hstart
initial bandwidth matrix, used in numerical optimisation

binned
flag for binned kernel estimation

bgridsize
vector of binning grid sizes

amise
flag to return the minimal scaled SCV value. Default is FALSE.

deriv.order
derivative order

verbose
flag to print out progress information. Default is FALSE.

optim.fun
optimiser function: one of nlm or optim

nstage
number of stages in the SCV bandwidth selector (1 or 2)

plot
flag to display plot of SCV(h) vs h (1-d only). Default is FALSE.

Details

hscv is the univariate SCV selector of Jones, Marron & Park (1991). Hscv is a multivariate generalisation of this, see Duong & Hazelton (2005). Use Hscv for unconstrained bandwidth matrices and Hscv.diag for diagonal bandwidth matrices.

The default pilot is "samse" for d=2, r=0, and "dscalar" otherwise. For SAMSE pilot bandwidths, see Duong & Hazelton (2005). Unconstrained and higher order derivative pilot bandwidths are from Chacon & Duong (2011).

For d=1, the selector hscv is not always stable for large sample sizes with binning. Examine the plot from hscv(plot=TRUE) to determine the appropriate smoothness of the SCV function. Any non-smoothness is due to the discretised nature of binned estimation.

For details about the advanced options for binned, Hstart, optim.fun, see Hpi.

Value

SCV bandwidth. If amise=TRUE then the minimal scaled SCV value is returned too.

References

Chacon, J.E. & Duong, T. (2011) Unconstrained pilot selectors for smoothed cross validation. Australian & New Zealand Journal of Statistics, 53, 331-351.

Duong, T. & Hazelton, M.L. (2005) Cross-validation bandwidth matrices for multivariate kernel density estimation. Scandinavian Journal of Statistics, 32, 485-506.

Jones, M.C., Marron, J.S. & Park, B.U. (1991) A simple root n bandwidth selector. Annals of Statistics, 19, 1919-1932.

See Also

Hbcv, Hlscv, Hpi

Examples

data(unicef)
Hscv(unicef)
hscv(unicef[,1])

[Package ks version 1.15.1 Index]