cardio {ks}	R Documentation

Foetal cardiotocograms

Description

This data set contains the cardiotocographic measurements from healthy, suspect and pathological foetuses.

Usage

data(cardio)

Format

A matrix with 2126 rows and 8 columns. Each row corresponds to a foetal cardiotocogram. The class label for the foetal state is the last column: N = normal, S = suspect, P = pathological. Details for all variables are found in the link below.

Source

Lichman, M. (2013) UCI Machine learning repository: cardiotocography data set. doi: 10.24432/C51S4N. University of California, Irvine, School of Information and Computer Sciences. Accessed 2017-05-18.

[Package ks version 1.15.1 Index]