kfe {ks}	R Documentation

Kernel functional estimate

Description

Kernel functional estimate for 1- to 6-dimensional data.

Usage

kfe(x, G, deriv.order, inc=1, binned, bin.par, bgridsize, deriv.vec=TRUE,
    add.index=TRUE, verbose=FALSE)
Hpi.kfe(x, nstage=2, pilot, pre="sphere", Hstart, binned=FALSE, 
    bgridsize, amise=FALSE, deriv.order=0, verbose=FALSE, optim.fun="optim")
Hpi.diag.kfe(x, nstage=2, pilot, pre="scale", Hstart, binned=FALSE,
    bgridsize, amise=FALSE, deriv.order=0, verbose=FALSE, optim.fun="optim")
hpi.kfe(x, nstage=2, binned=FALSE, bgridsize, amise=FALSE, deriv.order=0)

Arguments

x
vector/matrix of data values

nstage
number of stages in the plug-in bandwidth selector (1 or 2)

pilot
"dscalar" = single pilot bandwidth (default)
"dunconstr" = single unconstrained pilot bandwidth

pre
"scale" = pre.scale, "sphere" = pre.sphere

Hstart
initial bandwidth matrix, used in numerical optimisation

binned
flag for binned estimation

bgridsize
vector of binning grid sizes

amise
flag to return the minimal scaled PI value

deriv.order
derivative order

verbose
flag to print out progress information. Default is FALSE.

optim.fun
optimiser function: one of nlm or optim

G
pilot bandwidth matrix

inc
0=exclude diagonal, 1=include diagonal terms in kfe calculation

bin.par
binning parameters - output from binning

deriv.vec
flag to compute duplicated partial derivatives in the vectorised form. Default is FALSE.

add.index
flag to output derivative indices matrix. Default is true.

Details

Hpi.kfe is the optimal plug-in bandwidth for r-th order kernel functional estimator based on the unconstrained pilot selectors of Chacon & Duong (2010). hpi.kfe is the 1-d equivalent, using the formulas from Wand & Jones (1995, p.70).

kfe does not usually need to be called explicitly by the user.

Value

Plug-in bandwidth matrix for r-th order kernel functional estimator.

References

Chacon, J.E. & Duong, T. (2010) Multivariate plug-in bandwidth selection with unconstrained pilot matrices. Test, 19, 375-398.

Wand, M.P. & Jones, M.C. (1995) Kernel Smoothing. Chapman & Hall/CRC, London.

See Also

kde.test

[Package ks version 1.15.1 Index]