ksupp {ks}	R Documentation

Kernel support estimate

Description

Kernel support estimate for 2 and 3-dimensional data.

Usage

ksupp(fhat, cont=95, abs.cont, convex.hull=TRUE)

## S3 method for class 'ksupp'
plot(x, display="plot3D", ...)

Arguments

fhat
object of class kde

cont
percentage for contour level curve. Default is 95.

abs.cont
absolute density estimate height for contour level curve

convex.hull
flag to compute convex hull of contour level curve. Default is TRUE.

x
object of class ksupp

display
one of "plot3D", "rgl" (required for 3-d only)

...
other graphics parameters

Details

The kernel support estimate is the level set of the density estimate that exceeds the cont percent contour level. If this level set is a simply connected region, then this can suffice to be a conservative estimate of the density support. Otherwise, the convex hull of the level set is advised. For 2-d data, the convex hull is computed by chull; for 3-d data, it is computed by geometry::convhulln.

Value

A kernel support estimate is an object of class ksupp, i.e. a 2- or 3-column matrix which delimits the (convex hull of the) level set of the density estimate fhat.

See Also

kde

Examples

data(grevillea)
fhat <- kde(x=grevillea)
fhat.supp <- ksupp(fhat)
plot(fhat, display="filled.contour", cont=seq(10,90,by=10))
plot(fhat, cont=95, add=TRUE, col=1)
plot(fhat.supp, lty=2)

data(iris)
fhat <- kde(x=iris[,1:3])
fhat.supp <- ksupp(fhat)
plot(fhat)
plot(fhat.supp, add=TRUE, col=3, alpha=0.1)

[Package ks version 1.15.1 Index]